Updates and optimizations
- some updates necessary to use trilinos/develop, such as
passing an index to NgpField::get_component_stride(..)

- add m_syncCount and m_bucketIds in ExplicitTimeIntegrator to
facilitate using the version of stk::mesh::for_each_entity_run
that takes bucket-ids instead of a selector for improved
performance.

- added fences around some timers. it would be better to add
the fences inside the timer (fence before the timer is started
and before the timer is stopped). And probably use a flag or
macro so that the fences are not used in a release build or
when timers are not requested. The fences don't cause a huge
slowdown but probably a little...

- add a print in main to display the version of trilinos
and stk that are being used.